line_story_drf/requirements.txt
Django==4.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
drf-yasg==1.21.3
Pillow==9.2.0
psycopg2-binary==2.9.3
redis==4.3.4
